gadget: move TestOnDiskVolumeFromGadgetVol to the right place
