+++ gadget/ondisk_test.go
@@ -20,13 +20,16 @@
 package gadget_test
 
 import (
+	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"strings"
 
 	. "gopkg.in/check.v1"
 
 	"github.com/snapcore/snapd/gadget"
+	"github.com/snapcore/snapd/gadget/gadgettest"
 	"github.com/snapcore/snapd/gadget/quantity"
 	"github.com/snapcore/snapd/osutil/disks"
 	"github.com/snapcore/snapd/testutil"
@@ -416,4 +419,51 @@ func (s *ondiskTestSuite) TestOnDiskStructureFromPartition(c *C) {
 		PartitionFSType:  "ext4",
 		StartOffset:      1024 * 1024,
 	})
+}
+
+func (s *ondiskTestSuite) TestOnDiskVolumeFromGadgetVol(c *C) {
+	vdaSysPath := "/sys/devices/pci0000:00/0000:00:03.0/virtio1/block/vda"
+	restore := gadget.MockSysfsPathForBlockDevice(func(device string) (string, error) {
+		if strings.HasPrefix(device, "/dev/vda") == true {
+			return filepath.Join(vdaSysPath, filepath.Base(device)), nil
+		}
+		return "", fmt.Errorf("bad disk")
+	})
+	defer restore()
+
+	gadgetRoot := filepath.Join(c.MkDir(), "gadget")
+	ginfo, _, _, restore, err := gadgettest.MockGadgetPartitionedDisk(gadgettest.SingleVolumeClassicWithModesGadgetYaml, gadgetRoot)
+	c.Assert(err, IsNil)
+	defer restore()
+
+	// Initially without setting devices
+	diskVol, err := gadget.OnDiskVolumeFromGadgetVol(ginfo.Volumes["pc"])
+	c.Check(err.Error(), Equals, `volume "pc" has no device assigned`)
+	c.Check(diskVol, IsNil)
+
+	// Set devices as an installer would
+	for _, vol := range ginfo.Volumes {
+		for sIdx := range vol.Structure {
+			if vol.Structure[sIdx].Type == "mbr" {
+				continue
+			}
+			vol.Structure[sIdx].Device = fmt.Sprintf("/dev/vda%d", sIdx+1)
+		}
+	}
+
+	diskVol, err = gadget.OnDiskVolumeFromGadgetVol(ginfo.Volumes["pc"])
+	c.Check(err, IsNil)
+	c.Check(diskVol, DeepEquals, &gadgettest.MockGadgetPartitionedOnDiskVolume)
+
+	// Now setting it for the mbr
+	ginfo.Volumes["pc"].Structure[0].Device = "/dev/vda"
+	diskVol, err = gadget.OnDiskVolumeFromGadgetVol(ginfo.Volumes["pc"])
+	c.Check(err, IsNil)
+	c.Check(diskVol, DeepEquals, &gadgettest.MockGadgetPartitionedOnDiskVolume)
+
+	// Setting a wrong partition name
+	ginfo.Volumes["pc"].Structure[1].Device = "/dev/mmcblk0p1"
+	diskVol, err = gadget.OnDiskVolumeFromGadgetVol(ginfo.Volumes["pc"])
+	c.Check(err.Error(), Equals, "bad disk")
+	c.Check(diskVol, IsNil)
 }
